crud de reservas concluido

+++ app/sevices/__init__.py
@@ -11,8 +11,8 @@ def get_resources(resource: Reserva | Agendamento, db: Session, detail: str = 'N
     return db_resource if db_resource else not_found(detail)
 
 
-def get(resource: Reserva | Agendamento, db: Session, user_email: str, detail: str = 'Not found'):
-    db_resource = db.scalar(select(resource).where(resource.email_usuario == user_email))
+def get(resource: Reserva | Agendamento, db: Session, resource_id: int, detail: str = 'Not found'):
+    db_resource = db.scalar(select(resource).where(resource.id == resource_id))
 
     return db_resource if db_resource else not_found(detail)
 
@@ -35,10 +35,15 @@ def post(resource: Reserva | Agendamento, create_schema, db: Session):
     return db_resource
 
 
-def put(update_schema, user_email: str, resource: Reserva | Agendamento, db: Session, detail):
-    db_resource = get(resource=resource, user_email=user_email, db=db)
+def put(update_schema, resource_id:int, resource: Reserva | Agendamento, db: Session, detail):
+    db_resource = get(resource=resource, resource_id=resource_id, db=db)
 
-    update_data = update_schema.model_dump(exclude_unset=True)
+    if db_resource.email_usuario != update_schema.email_usuario:
+        return unauthorized(
+            detail='Você não tem permissão para atualizar esse recurso.'
+        )
+
+    update_data = update_schema.model_dump(exclude=['email_usuario'])
 
     for key, value in update_data.items():
         setattr(db_resource, key, value)
@@ -48,8 +53,13 @@ def put(update_schema, user_email: str, resource: Reserva | Agendamento, db: Ses
     return db_resource
 
 
-def delete(user_email: str, resource, db: Session, detail: str):
-    db_resource = get(resource=resource, user_email=user_email, db=db, detail=detail)
+def delete(delete_schema, resource_id:int, resource, db: Session, detail: str):
+    db_resource = get(resource=resource, resource_id=resource_id, db=db, detail=detail)
+
+    if db_resource.email_usuario != delete_schema.email_usuario:
+        return unauthorized(
+            detail="Você não tem permissão para deletar esse recurso."
+        )
 
     if db_resource:
         db.delete(db_resource)
